requirements.txt
python-telegram-bot==13.14
requests
beautifulsoup4
lxml
Pillow
urllib3==1.26.16
certifi
